Refactor Activitys, add Splash Activity
