+++ MarvelFanApp/app/build.gradle
@@ -44,4 +44,7 @@ dependencies {
     implementation 'com.android.support:recyclerview-v7:27.1.+'
 
     implementation 'com.github.bumptech.glide:glide:4.7.1'
+    implementation 'com.github.bumptech.glide:annotations:4.7.1'
+
+    annotationProcessor 'com.github.bumptech.glide:compiler:4.7.1'
 }
