+++ MarvelFanApp/app/src/main/java/com/example/oneshot/marvelfanapp/activity/SplashActivity.java
@@ -1,0 +1,31 @@
+package com.example.oneshot.marvelfanapp.activity;
+
+import android.app.Activity;
+import android.content.Intent;
+import android.os.Bundle;
+import android.os.Handler;
+import android.support.annotation.Nullable;
+import android.view.Window;
+
+import com.example.oneshot.marvelfanapp.MainActivity;
+import com.example.oneshot.marvelfanapp.R;
+
+public class SplashActivity extends Activity {
+
+    private final int DURATION_SPLASH = 3000;
+
+    @Override
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        this.requestWindowFeature(Window.FEATURE_NO_TITLE);
+        setContentView(R.layout.activity_splash);
+
+        new Handler().postDelayed(new Runnable(){
+            public void run(){
+                Intent intent = new Intent(SplashActivity.this, MainActivity.class);
+                startActivity(intent);
+                finish();
+            };
+        }, DURATION_SPLASH);
+    }
+}
