+++ MarvelFanApp/app/src/main/res/layout/activity_characters.xml
@@ -10,7 +10,7 @@
         android:layout_height="match_parent"
         android:layout_alignParentTop="true"
         android:scaleType="centerCrop"
-        android:src="@drawable/background"></ImageView>
+        android:src="@drawable/backgroundp"></ImageView>
 
 
     <ImageView
