+++ MarvelFanApp/app/src/main/res/layout/toolbar_character.xml
@@ -14,4 +14,4 @@
     app:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar"
 
 
-    tools:showIn="@layout/activity_main" />
+    tools:showIn="@layout/activity_list_characters" />
